+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 - Import ARB from Google Sheets
 
 ## 0.0.3
+
+- Generate dart class from arb file
+
+## 0.0.4
+
+- Server-side (private key) authentication